New: Added TravisCI Integration

+++ build.sh
@@ -14,10 +14,22 @@ SOURCE_DIR="source"
 SAMPLES_DIR="samples"
 DRIVER_FILE=${SOURCE_DIR}"/driver.xml"
 
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+PURPLE='\033[0;35m'
+NORM='\033[0m'  # No Color
+
 all()
 {
+    if [ ! -f $TEI_TO_RELAXNG_BIN ]; then
+        echo $TEI_TO_RELAXNG_BIN
+        echo "The TEI Stylesheets were not found at:" $PATH_TO_TEI_STYLESHEETS
+        echo "Please specify using -t flag."
+        exit 1
+    fi
+
     if [ -d "build" ]; then
-        echo "Removing old build directory"
+        echo -e "${PURPLE}Removing old build directory${NORM}"
         rm -r ${BUILD_DIR}
     fi
 
@@ -28,12 +40,12 @@ all()
 
     for file in $(find ${CUSTOMIZATIONS_DIR} -name '*.xml');
     do
-        echo "processing" "${file}"
-        $TEI_TO_RELAXNG_BIN --saxonjar=$PATH_TO_SAXON_JAR --trangjar=$PATH_TO_TRANG_JAR --localsource=$DRIVER_FILE $file $BUILD_DIR/$(basename ${file%%.*}).rng
+        echo -e "${GREEN}Processing" "${file}"$NORM
+        $TEI_TO_RELAXNG_BIN --localsource=$DRIVER_FILE $file $BUILD_DIR/$(basename ${file%%.*}).rng
 
         if [ $? = 1 ]; then
             IFS=$SAVEIFS
-            echo "Build failed on" $file
+            echo -e "${RED}Build failed on" $file$NORM
             exit 1
         fi
 
@@ -44,32 +56,57 @@ all()
 
 test()
 {
-    echo "Validating 2013 samples directory against mei-all"
+    echo -e "\nValidating 2013 samples directory against mei-all\n"
 
     SAVEIFS=$IFS
     IFS=$(echo -en "\n\b")
 
     for file in $(find ${SAMPLES_DIR}/MEI2013 -name '*.mei');
     do
-        echo "Testing: " $file
+        echo -e "${PURPLE} Testing: ${NORM}" $file
         $PATH_TO_JING $BUILD_DIR/mei-all.rng "${file}"
 
         if [ $? = 1 ]; then
             IFS=$SAVEIFS
-            echo "Tests failed on" $file
+            echo -e "${RED}\tTests failed on" $file$NORM
             exit 1
+        else
+            echo -e $GREEN '\t' $file "is valid against mei-all.rng${NORM}"
         fi
     done
 
     IFS=$SAVEIFS
 }
 
-case "$1" in
+usage()
+{
+    echo "Flags:"
+    echo "  -h Print usage"
+    echo "  -t Path to TEI Stylesheets"
+    echo ""
+    echo "Build options:"
+    echo "  all"
+    echo "  test"
+    exit 1
+}
+
+SKIP=0
+while getopts "h:t:" OPT; do
+    case $OPT in
+        h) 
+            usage;;
+        t)
+            PATH_TO_TEI_STYLESHEETS=$OPTARG
+            TEI_TO_RELAXNG_BIN="${PATH_TO_TEI_STYLESHEETS}/bin/teitorelaxng"
+            SKIP=$(($SKIP + 2));;
+    esac
+done
+
+args=("$@")
+TYPE=${args[$SKIP]}
+
+case $TYPE in
     "all" ) all;;
     "test" ) test;;
-    * )
-        echo "Build options:"
-        echo "  all"
-        echo "  test"
-    ;;
+    * ) usage;;
 esac
